+++ gen/aws/properties_definitions.go
@@ -156,6 +156,7 @@ var PropertiesDefinitions = []property{
 	{AwlessLabel: "PathPrefix", RDFLabel: fmt.Sprintf("%s:pathPrefix", rdf.CloudNS), RDFType: rdf.RdfProperty, RdfsDefinedBy: rdf.RdfsLiteral, RdfsDataType: rdf.XsdString},
 	{AwlessLabel: "PendingTasksCount", RDFLabel: fmt.Sprintf("%s:pendingTasksCount", rdf.CloudNS), RDFType: rdf.RdfProperty, RdfsDefinedBy: rdf.RdfsLiteral, RdfsDataType: rdf.XsdInt},
 	{AwlessLabel: "PlacementGroup", RDFLabel: fmt.Sprintf("%s:placementGroup", rdf.CloudNS), RDFType: rdf.RdfProperty, RdfsDefinedBy: rdf.RdfsLiteral, RdfsDataType: rdf.XsdString},
+	{AwlessLabel: "PlatformVersion", RDFLabel: fmt.Sprintf("%s:platformVersion", rdf.CloudNS), RDFType: rdf.RdfProperty, RdfsDefinedBy: rdf.RdfsLiteral, RdfsDataType: rdf.XsdString},
 	{AwlessLabel: "Port", RDFLabel: fmt.Sprintf("%s:port", rdf.NetNS), RDFType: rdf.RdfProperty, RdfsDefinedBy: rdf.RdfsLiteral, RdfsDataType: rdf.XsdInt},
 	{AwlessLabel: "Ports", RDFLabel: fmt.Sprintf("%s:ports", rdf.NetNS), RDFType: rdf.RdfProperty, RdfsDefinedBy: rdf.RdfsList, RdfsDataType: rdf.XsdString},
 	{AwlessLabel: "PortRange", RDFLabel: fmt.Sprintf("%s:portRange", rdf.NetNS), RDFType: rdf.RdfsSubProperty, RdfsDefinedBy: rdf.RdfsLiteral, RdfsDataType: rdf.XsdString},